@@ -1,20 +1,29 @@
 name: flutter_iconpicker
 description: A Dialog for picking Icons in Flutter and use them anywhere. Can be
   used as a default Dialog or as a Adaptive Dialog.
-version: 3.2.4
+version: 4.0.1
 homepage: https://github.com/Ahmadre
 repository: https://github.com/Ahmadre/FlutterIconPicker
+
 environment:
-  sdk: ^3.0.0
+  sdk: ">=3.3.4 <4.0.0"
+
 dependencies:
-  cupertino_icons: ^1.0.5
+  args: ^2.5.0
+  collection: ^1.18.0
+  cupertino_icons: ^1.0.8
+  dcli: ^6.0.5
+  equatable: ^2.0.5
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^10.2.1
-  provider: ^6.0.4
+  font_awesome_flutter: ^10.7.0
+  path: ^1.9.0
+  provider: ^6.1.2
+  scrollview_observer: ^1.21.2
+
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  flutter_lints: ">=4.0.0 <6.0.0"
+
 flutter:
   uses-material-design: true
   fonts:
