@@ -1,3 +1,3 @@
 import 'package:flutter_iconpicker/Models/icon_picker_icon.dart';
 
-Map<String, IconPickerIcon> allIcons = {};
+Map<String, IconPickerIcon> outlinedIcons = {};
