@@ -1,19 +1,48 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_iconpicker/Models/icon_pack.dart';
+import 'package:flutter_iconpicker/Models/icon_picker_icon.dart';
 
-const Map<String, IconData> myCustomIcons = {
-  'threesixty': Icons.threesixty,
-  'threed_rotation': Icons.threed_rotation,
-  'four_k': Icons.four_k,
-  'ac_unit': Icons.ac_unit,
-  'access_alarm': Icons.access_alarm,
-  'access_alarms': Icons.access_alarms,
-  'access_time': Icons.access_time,
-  'accessibility': Icons.accessibility,
-  'accessibility_new': Icons.accessibility_new,
-  'accessible': Icons.accessible,
-  'accessible_forward': Icons.accessible_forward,
-  'account_balance': Icons.account_balance,
-  'account_balance_wallet': Icons.account_balance_wallet,
-  'account_box': Icons.account_box,
-  'account_circle': Icons.account_circle,
+const Map<String, IconPickerIcon> myCustomIcons = {
+  'threesixty': IconPickerIcon(
+      name: 'threesixty', data: Icons.threesixty, pack: IconPack.custom),
+  'threed_rotation': IconPickerIcon(
+      name: 'threed_rotation',
+      data: Icons.threed_rotation,
+      pack: IconPack.custom),
+  'four_k':
+      IconPickerIcon(name: 'four_k', data: Icons.four_k, pack: IconPack.custom),
+  'ac_unit': IconPickerIcon(
+      name: 'ac_unit', data: Icons.ac_unit, pack: IconPack.custom),
+  'access_alarm': IconPickerIcon(
+      name: 'access_alarm', data: Icons.access_alarm, pack: IconPack.custom),
+  'access_alarms': IconPickerIcon(
+      name: 'access_alarms', data: Icons.access_alarms, pack: IconPack.custom),
+  'access_time': IconPickerIcon(
+      name: 'access_time', data: Icons.access_time, pack: IconPack.custom),
+  'accessibility': IconPickerIcon(
+      name: 'accessibility', data: Icons.accessibility, pack: IconPack.custom),
+  'accessibility_new': IconPickerIcon(
+      name: 'accessibility_new',
+      data: Icons.accessibility_new,
+      pack: IconPack.custom),
+  'accessible': IconPickerIcon(
+      name: 'accessible', data: Icons.accessible, pack: IconPack.custom),
+  'accessible_forward': IconPickerIcon(
+      name: 'accessible_forward',
+      data: Icons.accessible_forward,
+      pack: IconPack.custom),
+  'account_balance': IconPickerIcon(
+      name: 'account_balance',
+      data: Icons.account_balance,
+      pack: IconPack.custom),
+  'account_balance_wallet': IconPickerIcon(
+      name: 'account_balance_wallet',
+      data: Icons.account_balance_wallet,
+      pack: IconPack.custom),
+  'account_box': IconPickerIcon(
+      name: 'account_box', data: Icons.account_box, pack: IconPack.custom),
+  'account_circle': IconPickerIcon(
+      name: 'account_circle',
+      data: Icons.account_circle,
+      pack: IconPack.custom),
 };
