@@ -1,13 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_iconpicker/controllers/icon_controller.dart';
-import '../Helpers/ColorBrightness.dart';
-import '../IconPicker/iconPicker.dart';
-import '../IconPicker/searchBar.dart';
-import '../Models/IconPack.dart';
+import '../Helpers/color_brightness.dart';
+import '../IconPicker/icon_picker.dart';
+import '../IconPicker/multiple_icon_picker.dart';
+import '../IconPicker/search_bar.dart';
+import '../Models/icon_pack.dart';
+import '../Models/icon_picker_icon.dart';
 
 class FIPFullScreenDialog extends StatelessWidget {
   const FIPFullScreenDialog({
-    Key? key,
+    super.key,
     required this.iconController,
     required this.showSearchBar,
     required this.showTooltips,
@@ -17,23 +19,27 @@ class FIPFullScreenDialog extends StatelessWidget {
     required this.customIconPack,
     required this.searchIcon,
     required this.searchClearIcon,
+    required this.searchComparator,
     required this.searchHintText,
     required this.iconColor,
     required this.noResultsText,
     required this.iconSize,
     required this.mainAxisSpacing,
     required this.crossAxisSpacing,
-  }) : super(key: key);
+    this.selectedIconBackgroundColor,
+  });
 
   final FIPIconController iconController;
+  final Color? selectedIconBackgroundColor;
   final bool? showSearchBar;
   final bool? showTooltips;
   final Color? backgroundColor;
   final Widget? title;
   final List<IconPack>? iconPackMode;
-  final Map<String, IconData>? customIconPack;
+  final Map<String, IconPickerIcon>? customIconPack;
   final Icon? searchIcon;
   final Icon? searchClearIcon;
+  final SearchComparator? searchComparator;
   final String? searchHintText;
   final Color? iconColor;
   final String? noResultsText;
@@ -47,7 +53,7 @@ class FIPFullScreenDialog extends StatelessWidget {
       backgroundColor: backgroundColor,
       body: SafeArea(
         child: Padding(
-          padding: EdgeInsets.only(
+          padding: const EdgeInsets.only(
             top: 10,
             bottom: 20,
             left: 20,
@@ -55,21 +61,21 @@ class FIPFullScreenDialog extends StatelessWidget {
           ),
           child: Column(
             children: <Widget>[
-              Container(
+              SizedBox(
                 height: kToolbarHeight,
                 child: Row(
                   mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
                     Padding(
-                      padding: EdgeInsets.only(left: 6),
+                      padding: const EdgeInsets.only(left: 6),
                       child: DefaultTextStyle(
-                        child: title!,
                         style: TextStyle(
                           color: FIPColorBrightness(backgroundColor!).isLight()
                               ? Colors.black
                               : Colors.white,
                           fontSize: 20,
                         ),
+                        child: title!,
                       ),
                     ),
                     IconButton(
@@ -93,20 +99,38 @@ class FIPFullScreenDialog extends StatelessWidget {
                   searchClearIcon: searchClearIcon,
                   searchHintText: searchHintText,
                   backgroundColor: backgroundColor,
+                  searchComparator: searchComparator,
                 ),
               Expanded(
-                child: FIPIconPicker(
-                  iconController: iconController,
-                  showTooltips: showTooltips,
-                  iconPack: iconPackMode,
-                  customIconPack: customIconPack,
-                  iconColor: iconColor,
-                  backgroundColor: backgroundColor,
-                  noResultsText: noResultsText,
-                  iconSize: iconSize,
-                  mainAxisSpacing: mainAxisSpacing,
-                  crossAxisSpacing: crossAxisSpacing,
-                ),
+                child: iconController.isMultiple
+                    ? FIPMultipleIconPicker(
+                        iconController: iconController,
+                        selectedIconBackgroundColor:
+                            selectedIconBackgroundColor,
+                        showTooltips: showTooltips,
+                        iconPack: iconPackMode,
+                        customIconPack: customIconPack,
+                        iconColor: iconColor,
+                        backgroundColor: backgroundColor,
+                        noResultsText: noResultsText,
+                        iconSize: iconSize,
+                        mainAxisSpacing: mainAxisSpacing,
+                        crossAxisSpacing: crossAxisSpacing,
+                      )
+                    : FIPIconPicker(
+                        iconController: iconController,
+                        selectedIconBackgroundColor:
+                            selectedIconBackgroundColor,
+                        showTooltips: showTooltips,
+                        iconPack: iconPackMode,
+                        customIconPack: customIconPack,
+                        iconColor: iconColor,
+                        backgroundColor: backgroundColor,
+                        noResultsText: noResultsText,
+                        iconSize: iconSize,
+                        mainAxisSpacing: mainAxisSpacing,
+                        crossAxisSpacing: crossAxisSpacing,
+                      ),
               ),
             ],
           ),
