@@ -8,20 +8,32 @@ export 'Packs/Cupertino.dart';
 export 'Packs/FontAwesome.dart';
 export 'Packs/LineIcons.dart';
 
-import '../Models/IconPack.dart';
+import 'package:flutter_iconpicker/IconPicker/Packs/MaterialDefault.dart';
+import 'package:flutter_iconpicker/IconPicker/Packs/MaterialOutlined.dart';
+import 'package:flutter_iconpicker/IconPicker/Packs/MaterialRounded.dart';
+import 'package:flutter_iconpicker/IconPicker/Packs/MaterialSharp.dart';
 
+import '../Models/icon_pack.dart';
+
+import '../Models/icon_picker_icon.dart';
 import 'Packs/Material.dart';
 import 'Packs/Cupertino.dart';
 import 'Packs/FontAwesome.dart';
 import 'Packs/LineIcons.dart';
 
-import 'package:flutter/widgets.dart';
-
 class FIPIconManager {
-  static Map<String, IconData> getSelectedPack(IconPack? pickedPack) {
+  static Map<String, IconPickerIcon> getSelectedPack(IconPack? pickedPack) {
     switch (pickedPack) {
       case IconPack.material:
-        return icons;
+        return defaultIcons;
+      case IconPack.allMaterial:
+        return allIcons;
+      case IconPack.sharpMaterial:
+        return sharpIcons;
+      case IconPack.roundedMaterial:
+        return roundedIcons;
+      case IconPack.outlinedMaterial:
+        return outlinedIcons;
       case IconPack.cupertino:
         return cupertinoIcons;
       case IconPack.fontAwesomeIcons:
@@ -29,7 +41,7 @@ class FIPIconManager {
       case IconPack.lineAwesomeIcons:
         return lineAwesomeIcons;
       default:
-        return <String, IconData>{};
+        return <String, IconPickerIcon>{};
     }
   }
 }
